Extends Svelte stores to implements JS object more easily
Svelte stores are not very easy to use with JS objects such as Map, Set
and Array.
Instead of calling simple methods (such as array.push(42) ), we have to
update the store and return the reference of the object (such as
storeArray.update(arr => { arr.push(42); return arr; }) ).

This is a lot of boilerplate and is quite hard to read.

This is why, SvelteStore helper has been created, to build some built-in
store of JS objects.
Now with these store, it is possible to call JS methods directly on the
store.

Example:
  const list = writableArray([1, 2]);
  list.push(3); // updates the store with reactivity
  list.at(-1); // 3

list[2]; is not possible because "list" is still a Svelte store.

It works for Map, and Set.
For Array, only some methods have been implemented, and this should be
completed if needed.