Do not use achievement to ends battle
achievement are triggered once and stay active over runs. But it should
be reset from one run to another.

using equipment is much better, moreover when equipments are not
displayed in Tutorial story.

The "done" action may now be triggered only when the action is finished.
This avoids triggering this event too early and display wrong logs.